+++ src/main/java/mc/sloth/slothplugin/commands/Media/DiscordCommand.java
@@ -1,25 +1,18 @@
-package mc.sloth.slothplugin.commands;
+package mc.sloth.slothplugin.commands.Media;
 
 import mc.sloth.slothplugin.files.config.CustomConfig;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 
-public class TntCommand implements CommandExecutor {
+public class DiscordCommand implements CommandExecutor {
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         if (sender instanceof Player) {
             Player player = (Player) sender;
-
-            if (player.hasPermission("slothplugin.tnt")) {
-                player.getWorld().spawnEntity(player.getLocation(), EntityType.PRIMED_TNT);
-                player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("TntCommandMessage")));
-            } else {
-                player.sendMessage(ChatColor.RED + CustomConfig.get().getString("MissingPermissionsMessage"));
-            }
+            player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("Discord")));
         } else {
             System.out.println("Ta komende mozesz wywolac tylko w grze");
         }