+++ src/main/java/mc/sloth/slothplugin/commands/Media/TeamspeakCommand.java
@@ -1,4 +1,4 @@
-package mc.sloth.slothplugin.commands;
+package mc.sloth.slothplugin.commands.Media;
 
 import mc.sloth.slothplugin.files.config.CustomConfig;
 import org.bukkit.ChatColor;
@@ -7,17 +7,12 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-public class StrikeCommand implements CommandExecutor {
+public class TeamspeakCommand implements CommandExecutor {
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         if (sender instanceof Player) {
             Player player = (Player) sender;
-            if (player.hasPermission("slothplugin.strike")) {
-                player.getWorld().strikeLightning(player.getLocation());
-                player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("StrikeCommandMessage")));
-            } else {
-                player.sendMessage(ChatColor.RED + CustomConfig.get().getString("MissingPermissionsMessage"));
-            }
+            player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("TeamSpeak")));
         } else {
             System.out.println("Ta komende mozesz wywolac tylko w grze");
         }