+++ src/main/java/mc/sloth/slothplugin/commands/Media/YoutubeCommand.java
@@ -1,4 +1,4 @@
-package mc.sloth.slothplugin.commands;
+package mc.sloth.slothplugin.commands.Media;
 
 import mc.sloth.slothplugin.files.config.CustomConfig;
 import org.bukkit.ChatColor;
@@ -7,17 +7,12 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-public class FeedCommand implements CommandExecutor {
+public class YoutubeCommand implements CommandExecutor {
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         if (sender instanceof Player) {
             Player player = (Player) sender;
-            if (player.hasPermission("slothplugin.feed")) {
-                player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("FeedCommandMessage")));
-                player.setFoodLevel(20);
-            } else {
-                player.sendMessage(ChatColor.RED + CustomConfig.get().getString("MissingPermissionsMessage"));
-            }
+            player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("YoutubeChannel")));
         } else {
             System.out.println("Ta komende mozesz wywolac tylko w grze");
         }