+++ src/main/java/mc/sloth/slothplugin/commands/Player/FeedCommand.java
@@ -1,4 +1,4 @@
-package mc.sloth.slothplugin.commands;
+package mc.sloth.slothplugin.commands.Player;
 
 import mc.sloth.slothplugin.files.config.CustomConfig;
 import org.bukkit.ChatColor;
@@ -7,19 +7,14 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-public class FlyCommand implements CommandExecutor {
+public class FeedCommand implements CommandExecutor {
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         if (sender instanceof Player) {
             Player player = (Player) sender;
-            if (player.hasPermission("slothplugin.fly")) {
-                if (player.getAllowFlight()) {
-                    player.setAllowFlight(false);
-                    player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("FlyOffCommandMessage")));
-                } else {
-                    player.setAllowFlight(true);
-                    player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("FlyOnCommandMessage")));
-                }
+            if (player.hasPermission("slothplugin.feed")) {
+                player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("FeedCommandMessage")));
+                player.setFoodLevel(20);
             } else {
                 player.sendMessage(ChatColor.RED + CustomConfig.get().getString("MissingPermissionsMessage"));
             }