+++ src/main/java/mc/sloth/slothplugin/commands/Player/HealCommand.java
@@ -1,4 +1,4 @@
-package mc.sloth.slothplugin.commands;
+package mc.sloth.slothplugin.commands.Player;
 
 import mc.sloth.slothplugin.files.config.CustomConfig;
 import org.bukkit.ChatColor;
@@ -7,20 +7,22 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-public class CraftingCommand implements CommandExecutor {
+public class HealCommand implements CommandExecutor {
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         if (sender instanceof Player) {
             Player player = (Player) sender;
-            if (player.hasPermission("slothplugin.crafting")) {
-                player.openWorkbench(null, true);
-                player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("CraftingCommandMessage")));
+            if (player.hasPermission("slothplugin.heal")) {
+                player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("HealCommandMessage")));
+                player.setHealth(20.0);
+                player.setFoodLevel(20);
             } else {
                 player.sendMessage(ChatColor.RED + CustomConfig.get().getString("MissingPermissionsMessage"));
             }
         } else {
             System.out.println("Ta komende mozesz wywolac tylko w grze");
         }
+
         return true;
     }
 }