+++ src/main/java/mc/sloth/slothplugin/commands/World/StrikeCommand.java
@@ -1,0 +1,26 @@
+package mc.sloth.slothplugin.commands.World;
+
+import mc.sloth.slothplugin.files.config.CustomConfig;
+import org.bukkit.ChatColor;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
+public class StrikeCommand implements CommandExecutor {
+    @Override
+    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+        if (sender instanceof Player) {
+            Player player = (Player) sender;
+            if (player.hasPermission("slothplugin.strike")) {
+                player.getWorld().strikeLightning(player.getLocation());
+                player.sendMessage(ChatColor.translateAlternateColorCodes('&', CustomConfig.get().getString("StrikeCommandMessage")));
+            } else {
+                player.sendMessage(ChatColor.RED + CustomConfig.get().getString("MissingPermissionsMessage"));
+            }
+        } else {
+            System.out.println("Ta komende mozesz wywolac tylko w grze");
+        }
+        return true;
+    }
+}